Empty types in parameters in generated swagger schema json file
## Type
- [X] Bug
- [ ] Feature
- [ ] Refactoring
- [X] Performance improvement
- [ ] Integration test

## Problem setting

Some parameter types are defined as nullable in the form "type | undefined" instead of "type?"

In this format, swagger does not recognize the type in the json schema file it generates in the /api-json path, leaving the "schema" node empty

In addition, the optional pipes, by returning undefined, also mask the data type.

## Proposed Changes

Replace the definition of nullable types of the form "type | undefined" to "type?" and return string instead undefined in optional pipes that doesn't affect to the pipe.

## How to test

Get the json schema file from swagger in the path /api-json and check, for example, in "blocks/count" the parameters "shard", "proposer", etc...